Fix Bot Not trading items less than ignore rate
Currently if your bots inventory has a item that's price is less than the site's ignore rate, it will get its value set to zero just like it would for the user. This fix lets users trade for an item from the bot that is worth less than the ignore rate while still not  allowing the user to trade the site items that have a price less than your ignore rate.
For example:
The bot has a item worth .49 cents, your ignore rate is set to ignore all skins less than .50 cents. When a user would try to trade for the skin that is worth less than the ignore rate they would get the error "Could not get a price for item(s). Trade has been cancelled". This is because the bot was using the same getprice function as the user would and it would set the price too zero because it is less than the ignore rate.