Fixes Price Not Being Shown Correctly On Front End
When an item would be classified as trash it's price would be shown wrong on the front end. This is because it would ONLY multiply the items price by the trash rate. In the back end the trash multiplier is added on TOP of the item type rate. This would lead to the user getting a 'not enough value' error on the front end because the items price would be higher on the backend than it was shown to the user on the front end.